requirements.txt
numpy
pandas
scikit-learn
stable-baselines3[extra]
langchain
pymupdf
requests
openai
python-dotenv
google-generativeai
langchain-google-genai
langchain-community
faiss-cpu
sentence-transformers
torch
gymnasium
pytesseract
pillow
numpy-stl
